requirements.txt
tensorflow==2.16.1
boto3==1.34.49
keras==2.12.0
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.1
dvc==2.18.0
dvc-s3==3.0.1
pylint==2.13
flake8==7.0.0
lib-ml-remla-team14-a2==1.0.14.dev0
typing-extensions==4.5.0
scmrepo==0.0.25
pydantic==1.9.0
pytest